Rename: pk 타입 변경

+++ backend/helpus_donation/helpus/src/main/java/com/ssafy/helpus/donation/controller/DonationController.java
@@ -81,7 +81,7 @@ public class DonationController {
 
     @ApiOperation(value = "기부 글 조회")
     @GetMapping("{donationId}")
-    public ResponseEntity getDonation(@PathVariable Integer donationId) {
+    public ResponseEntity getDonation(@PathVariable Long donationId) {
         log.info("DonationController getDonation call");
 
         Map<String, Object> resultMap = new HashMap<>();
@@ -99,7 +99,7 @@ public class DonationController {
 
     @ApiOperation(value = "기부 글 마감")
     @DeleteMapping("{donationId}")
-    public ResponseEntity endDonation(@PathVariable Integer donationId) {
+    public ResponseEntity endDonation(@PathVariable Long donationId) {
         log.info("DonationController endDonation call");
 
         Map<String, Object> resultMap = new HashMap<>();
@@ -118,7 +118,7 @@ public class DonationController {
     @ApiOperation(value = "기부 글 목록 조회")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "memberId", value = "작성자 고유 번호", required = false,
-                    dataType = "Integer", paramType = "query"),
+                    dataType = "Long", paramType = "query"),
             @ApiImplicitParam(name = "order", value = "정렬 순서", required = false,
                     dataType = "String", paramType = "query"),
             @ApiImplicitParam(name = "word", value = "검색어", required = false,
@@ -127,7 +127,7 @@ public class DonationController {
                     dataType = "int", paramType = "query")
     })
     @GetMapping
-    public ResponseEntity listDonation(@RequestParam(required = false) Integer memberId,
+    public ResponseEntity listDonation(@RequestParam(required = false) Long memberId,
                                        @RequestParam(required = false, defaultValue = "최신순") String order,
                                        @RequestParam(required = false, defaultValue = "1") int page) {
         log.info("DonationController listDonation call");
