+++ backend/helpus_donation/helpus/src/main/java/com/ssafy/helpus/donation/dto/Confirm/ConfirmUpdateReqDto.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.NotNull;
 public class ConfirmUpdateReqDto {
     @ApiModelProperty(value = "후기 글 고유 번호")
     @NotNull(message = "후기 글 고유 번호를 입력해주세요")
-    private Integer donationConfirmId;
+    private Long donationConfirmId;
 
     @ApiModelProperty(value = "제목")
     @NotBlank(message = "제목을 입력해주세요")
